# hw2/openapi.yaml
openapi: 3.0.3
info:
  title: TableTurn API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
paths:
  /api/parties:
    get:
      summary: List parties in arrival order
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [waiting, seated, removed]
      responses:
        '200':
          description: Parties
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Party' }
    post:
      summary: Add a party to the waiting list
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewParty' }
      responses:
        '201':
          description: Created party
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Party' }
        '422': { description: Invalid name or party size }
  /api/parties/{party_id}:
    patch:
      summary: Seat or remove a waiting party
      parameters:
        - name: party_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [seated, removed] }
      responses:
        '200':
          description: Updated party
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Party' }
        '404': { description: Party not found }
        '409': { description: Party is no longer waiting }
components:
  schemas:
    NewParty:
      type: object
      required: [name, size]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        size: { type: integer, minimum: 1, maximum: 20 }
    Party:
      allOf:
        - $ref: '#/components/schemas/NewParty'
        - type: object
          required: [id, status, created_at]
          properties:
            id: { type: integer }
            status: { type: string, enum: [waiting, seated, removed] }
            created_at: { type: string, format: date-time }